requirements.txt
telethon
fastapi
uvicorn
psycopg2-binary
dbt-postgres
ultralytics
dagster
python-dotenv
pydantic
sqlalchemy
httpx
dotenv
dbt-postgres